requirements.txt
numpy==1.16.4
scipy==1.3.1
tensorflow-gpu==1.14.0
Pillow
matplotlib
jupyter
pandas==0.24.2
timeout-decorator
eventlet
